X509: PKI code refactoring (clippy, doc)
PKI refactoring

Clippy improvements
