+++ keysas_lib/tests/pki_test.rs
@@ -2,8 +2,8 @@ use ed25519_dalek::Digest;
 use ed25519_dalek::Keypair;
 use ed25519_dalek::Sha512;
 use hex_literal::hex;
-use keysas_lib::pki::KeysasKey;
-use keysas_lib::pki::KeysasPQKey;
+use keysas_lib::keysas_key::KeysasKey;
+use keysas_lib::keysas_key::KeysasPQKey;
 use oqs::sig::Algorithm;
 use oqs::sig::Sig;
 use pkcs8::der::Any;
